+++ src/main/java/com/ggapp/dao/repository/mysql/ProductVoucherRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductVoucherRepository extends JpaRepository<ProductVoucher, Long> {
+public interface ProductVoucherRepository extends JpaRepository<ProductVoucher, Integer> {
 }
