+++ src/main/java/com/ggapp/services/BrandService.java
@@ -10,8 +10,8 @@ public interface BrandService {
 	BrandResponse createBrand(BrandRequest brandRequest, CustomUserDetail customUserDetail) throws ApplicationException;
 	CommonResponsePayload getAllBrand(int page, int size);
 	CommonResponsePayload getBrandbyKeyword(int page, int size, String keyword);
-	BrandResponse updateBrand(Long id, BrandRequest brandRequest, CustomUserDetail customUserDetail) throws ApplicationException;
-	BrandResponse logicDeleteBrand(Long id, CustomUserDetail customUserDetail) throws ApplicationException;
-	boolean physicalDeleteBrand(Long id) throws ApplicationException;
+	BrandResponse updateBrand(int id, BrandRequest brandRequest, CustomUserDetail customUserDetail) throws ApplicationException;
+	BrandResponse logicDeleteBrand(int id, CustomUserDetail customUserDetail) throws ApplicationException;
+	boolean physicalDeleteBrand(int id) throws ApplicationException;
 	boolean isExists(String brandName);
 }
