+++ src/main/java/com/ggapp/services/CartService.java
@@ -7,12 +7,12 @@ import com.ggapp.common.exception.ApplicationException;
  * @author Tran Minh Truyen
  */
 public interface CartService {
-	CartResponse createCart(Long customerId, Long productId, Long storeId, long productAmount) throws ApplicationException;
-	CartResponse getCartById(Long id) throws ApplicationException;
-	CartResponse updateProductAmountInCart(Long customerId, Long productId, Long storeId, long amount) throws ApplicationException;
-	CartResponse deleteCart(Long id, Long storeId) throws ApplicationException;
-	boolean deleteCartAfterCreateOrder(Long id);
-	CartResponse addProductToCart(Long customerId, Long productId, Long storeId, long productAmount) throws ApplicationException;
-	CartResponse removeProductFromCart(Long customerId, Long productId, Long storeId) throws ApplicationException;
-	boolean isCartExists(Long customerId);
+	CartResponse createCart(int customerId, int productId, int storeId, int productAmount) throws ApplicationException;
+	CartResponse getCartById(int id) throws ApplicationException;
+	CartResponse updateProductAmountInCart(int customerId, int productId, int storeId, int amount) throws ApplicationException;
+	CartResponse deleteCart(int id, int storeId) throws ApplicationException;
+	boolean deleteCartAfterCreateOrder(int id);
+	CartResponse addProductToCart(int customerId, int productId, int storeId, int productAmount) throws ApplicationException;
+	CartResponse removeProductFromCart(int customerId, int productId, int storeId) throws ApplicationException;
+	boolean isCartExists(int customerId);
 }
