+++ src/main/java/com/ggapp/services/NewsService.java
@@ -7,9 +7,9 @@ import com.ggapp.common.dto.response.NewsResponse;
 import com.ggapp.common.exception.ApplicationException;
 
 public interface NewsService {
-	NewsResponse createNews (Long userCreateId, NewsRequest newsRequest) throws ApplicationException;
+	NewsResponse createNews (int userCreateId, NewsRequest newsRequest) throws ApplicationException;
 	CommonResponsePayload getAllNews (int page, int size) throws ApplicationException;
-	CommonResponsePayload getNewsByKeyWord (int page, int size, Long userCreateId, String userCreateName, String title) throws Exception;
-	NewsResponse updateNews (Long newsId, NewsRequest newsRequest) throws ApplicationException;
-	BaseResponse deleteNews (Long newsId) throws ApplicationException;
+	CommonResponsePayload getNewsByKeyWord (int page, int size, int userCreateId, String userCreateName, String title) throws Exception;
+	NewsResponse updateNews (int newsId, NewsRequest newsRequest) throws ApplicationException;
+	BaseResponse deleteNews (int newsId) throws ApplicationException;
 }
